Hoan tat flow xu li voi input m, ve do thi va truong hop binh thuong

+++ app/src/main/java/com/example/highschoolmathsolver/mathengine/expression/MonomialExp.kt
@@ -2,13 +2,13 @@ package com.example.highschoolmathsolver.mathengine.expression
 
 import kotlin.math.pow
 
-class MonomialExp(_a: Double, _n: Double) : FormalExpression {
+class MonomialExp(_a: Double, _n: Int) : FormalExpression {
     private var a: Double = _a
-    private var n: Double = _n
+    private var n: Int = _n
 
     companion object {
-        public fun creatInstance(_a: Double, _n: Double): FormalExpression {
-            if (_a == 0.0 || _n == 0.0) {
+        public fun creatInstance(_a: Double, _n: Int): FormalExpression {
+            if (_a == 0.0 || _n == 0) {
                 return ConstExp(_a)
             }
             return MonomialExp(_a, _n)
@@ -21,7 +21,7 @@ class MonomialExp(_a: Double, _n: Double) : FormalExpression {
 
     override fun derive(): FormalExpression {
         var aDerive: Double = this.a * this.n
-        var nDerive: Double = this.n - 1;
+        var nDerive: Int = this.n - 1;
         return MonomialExp.creatInstance(aDerive, nDerive)
     }
     override fun expToString(): String {
@@ -32,7 +32,7 @@ class MonomialExp(_a: Double, _n: Double) : FormalExpression {
         } else if(this.a==-1.0) {
             return "-x^"+this.n
         }
-        else if (this.n == 1.0) {
+        else if (this.n == 1) {
             return this.a.toString() + "x"
         } else {
             return this.a.toString() + "x^" + this.n.toString()
